Updated "bounce.xml" fixed xml bugg
I found that if you save an xml attribute in the mutationtodom or
domtomutation you have to use lowercase letters (no camel case) or it
will not load it correctly. I fixed all mutators in the shapes folder
to reflect this and put a comment about it.

I added drop down menus for ring and arrow and added them to the
“objectDropDown” list.

updated “bounce.xml” to comply with new “get” and “set” blocks