Added first version of vector math block
some thought needs to go into how to make it work more user friendly.

right now there isn't any prompt for how to use it. When a drop down selection that requires another vector is picked it attaches another input correctly and it currently generates the code correctly.

A couple of ideas for how it could be improved

	1. make the added input be inline and have the math operator
	   be inserted between the blocks for example

		ex. [block] X [block] for cross product
		    [block] · [block] for dot product