fixes #4 prevents set variable from being stuck when selecting a new variaable
also fixes a similar bug that wasn't documented with the get block

the code was checking "children" of the definition set block which
can be the block below it or its connected block. New code only looks
at connected block.

this commit also comments out "onChange" function as it is incrediable unstable
and is making all debugging difficult